// src/types/globalTypes.ts
export interface CollectionCreateInput {
    isPublished?: boolean | null;
    name?: string | null;
    slug?: string | null;
    description?: any | null;
    backgroundImage?: any | null;
    backgroundImageAlt?: string | null;
    seo?: SeoInput | null;
    publicationDate?: any | null;
    products?: (string | null)[] | null;
}
// ...
export interface SeoInput {
    title?: string | null;
    description?: string | null;
}